<==== Mercurial data ====>
Branch: master

+++ common/src/utils/color_space/yuvconvert.cpp
@@ -273,7 +273,7 @@ void yuv420_argb32_sse2_intr(unsigned char * dst, const unsigned char * py,
 
 void bgra_to_yv12_sse2_intr(const quint8* rgba, int xStride, quint8* y, quint8* u, quint8* v, int yStride, int uvStride, int width, int height, bool flip)
 {
-    Q_ASSERT( qPower2Ceil((unsigned int)xStride, 32) == xStride );
+    Q_ASSERT( qPower2Ceil((unsigned int)xStride, 32) == (unsigned int)xStride );
 
     static const __m128i sse_2000         = _mm_setr_epi16( 0x2020, 0x2020, 0x2020, 0x2020, 0x2020, 0x2020, 0x2020, 0x2020 ); /* SSE2. */
     static const __m128i sse_00a0         = _mm_setr_epi16( 0x0210, 0x0210, 0x0210, 0x0210, 0x0210, 0x0210, 0x0210, 0x0210 ); /* SSE2. */
